backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Gemini only - no other AI providers
langchain-google-genai==0.0.7
google-generativeai==0.3.2

# Core LangChain (minimal)
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21

# Vector store (FAISS only)
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0